+++ plugins/auth-backend/alpha-api-report.md
@@ -3,24 +3,19 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { AuthProviderExtensionPoint } from '@backstage/plugin-auth-node/alpha';
+import { authProviderExtensionPoint } from '@backstage/plugin-auth-node/alpha';
 import { BackendFeature } from '@backstage/backend-plugin-api';
-import { ExtensionPoint } from '@backstage/backend-plugin-api';
-import { ProviderFactories } from '@backstage/plugin-auth-backend';
+
+// @alpha
+export const authModuleDemoAuthProviders: () => BackendFeature;
 
 // @alpha
 export const authPlugin: () => BackendFeature;
 
-// @alpha (undocumented)
-export interface AuthProviderExtensionPoint {
-  // (undocumented)
-  addProviderFactories(factories: ProviderFactories): void;
-}
+export { AuthProviderExtensionPoint };
 
-// @alpha (undocumented)
-export const authProviderExtensionPoint: ExtensionPoint<AuthProviderExtensionPoint>;
-
-// @alpha
-export const defaultAuthProviders: () => BackendFeature;
+export { authProviderExtensionPoint };
 
 // (No @packageDocumentation comment for this package)
 ```
