+++ plugins/auth-node/alpha-api-report.md
@@ -1,0 +1,19 @@
+## API Report File for "@backstage/plugin-auth-node"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { AuthProviderFactory } from '@backstage/plugin-auth-node';
+import { ExtensionPoint } from '@backstage/backend-plugin-api';
+
+// @alpha (undocumented)
+export interface AuthProviderExtensionPoint {
+  // (undocumented)
+  addProviders(factories: Record<string, AuthProviderFactory>): void;
+}
+
+// @alpha (undocumented)
+export const authProviderExtensionPoint: ExtensionPoint<AuthProviderExtensionPoint>;
+
+// (No @packageDocumentation comment for this package)
+```
