+++ plugins/auth-node/api-report.md
@@ -3,8 +3,89 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { Config } from '@backstage/config';
+import { Entity } from '@backstage/catalog-model';
+import express from 'express';
+import { GetEntitiesRequest } from '@backstage/catalog-client';
+import { JsonValue } from '@backstage/types';
+import { Logger } from 'winston';
 import { PluginEndpointDiscovery } from '@backstage/backend-common';
+import { Profile } from 'passport';
 import { Request as Request_2 } from 'express';
+
+// @public
+export type AuthHandler<TAuthResult> = (
+  input: TAuthResult,
+  context: AuthResolverContext,
+) => Promise<AuthHandlerResult>;
+
+// @public
+export type AuthHandlerResult = {
+  profile: ProfileInfo;
+};
+
+// @public (undocumented)
+export type AuthProviderConfig = {
+  baseUrl: string;
+  appUrl: string;
+  isOriginAllowed: (origin: string) => boolean;
+  cookieConfigurer?: CookieConfigurer;
+};
+
+// @public (undocumented)
+export type AuthProviderFactory = (options: {
+  providerId: string;
+  globalConfig: AuthProviderConfig;
+  config: Config;
+  logger: Logger;
+  resolverContext: AuthResolverContext;
+}) => AuthProviderRouteHandlers;
+
+// @public
+export interface AuthProviderRouteHandlers {
+  frameHandler(req: express.Request, res: express.Response): Promise<void>;
+  logout?(req: express.Request, res: express.Response): Promise<void>;
+  refresh?(req: express.Request, res: express.Response): Promise<void>;
+  start(req: express.Request, res: express.Response): Promise<void>;
+}
+
+// @public
+export type AuthResolverCatalogUserQuery =
+  | {
+      entityRef:
+        | string
+        | {
+            kind?: string;
+            namespace?: string;
+            name: string;
+          };
+    }
+  | {
+      annotations: Record<string, string>;
+    }
+  | {
+      filter: Exclude<GetEntitiesRequest['filter'], undefined>;
+    };
+
+// @public
+export type AuthResolverContext = {
+  issueToken(params: TokenParams): Promise<{
+    token: string;
+  }>;
+  findCatalogUser(query: AuthResolverCatalogUserQuery): Promise<{
+    entity: Entity;
+  }>;
+  signInWithCatalogUser(
+    query: AuthResolverCatalogUserQuery,
+  ): Promise<BackstageSignInResult>;
+};
+
+// @public (undocumented)
+export type AuthResponse<ProviderInfo> = {
+  providerInfo: ProviderInfo;
+  profile: ProfileInfo;
+  backstageIdentity?: BackstageIdentityResponse;
+};
 
 // @public
 export interface BackstageIdentityResponse extends BackstageSignInResult {
@@ -21,6 +102,19 @@ export type BackstageUserIdentity = {
   type: 'user';
   userEntityRef: string;
   ownershipEntityRefs: string[];
+};
+
+// @public
+export type CookieConfigurer = (ctx: {
+  providerId: string;
+  baseUrl: string;
+  callbackUrl: string;
+  appOrigin: string;
+}) => {
+  domain: string;
+  path: string;
+  secure: boolean;
+  sameSite?: 'none' | 'lax' | 'strict';
 };
 
 // @public
@@ -64,5 +158,123 @@ export type IdentityClientOptions = {
   discovery: PluginEndpointDiscovery;
   issuer?: string;
   algorithms?: string[];
+};
+
+// @public
+export interface OAuthHandlers {
+  handler(req: express.Request): Promise<{
+    response: OAuthResponse;
+    refreshToken?: string;
+  }>;
+  logout?(req: OAuthLogoutRequest): Promise<void>;
+  refresh?(req: OAuthRefreshRequest): Promise<{
+    response: OAuthResponse;
+    refreshToken?: string;
+  }>;
+  start(req: OAuthStartRequest): Promise<OAuthStartResponse>;
+}
+
+// @public (undocumented)
+export type OAuthLogoutRequest = express.Request<{}> & {
+  refreshToken: string;
+};
+
+// @public (undocumented)
+export type OAuthProviderInfo = {
+  accessToken: string;
+  idToken?: string;
+  expiresInSeconds?: number;
+  scope: string;
+};
+
+// @public
+export type OAuthProviderOptions = {
+  clientId: string;
+  clientSecret: string;
+  callbackUrl: string;
+};
+
+// @public (undocumented)
+export type OAuthRefreshRequest = express.Request<{}> & {
+  scope: string;
+  refreshToken: string;
+};
+
+// @public
+export type OAuthResponse = {
+  profile: ProfileInfo;
+  providerInfo: OAuthProviderInfo;
+  backstageIdentity?: BackstageSignInResult;
+};
+
+// @public (undocumented)
+export type OAuthResult = {
+  fullProfile: Profile;
+  params: {
+    id_token?: string;
+    scope: string;
+    expires_in: number;
+  };
+  accessToken: string;
+  refreshToken?: string;
+};
+
+// @public (undocumented)
+export type OAuthStartRequest = express.Request<{}> & {
+  scope: string;
+  state: OAuthState;
+};
+
+// @public (undocumented)
+export type OAuthStartResponse = {
+  url: string;
+  status?: number;
+};
+
+// @public (undocumented)
+export type OAuthState = {
+  nonce: string;
+  env: string;
+  origin?: string;
+  scope?: string;
+  redirectUrl?: string;
+  flow?: string;
+};
+
+// @public
+export type ProfileInfo = {
+  email?: string;
+  displayName?: string;
+  picture?: string;
+};
+
+// @public (undocumented)
+export type ProviderFactories = {
+  [s: string]: AuthProviderFactory;
+};
+
+// @public
+export type SignInInfo<TAuthResult> = {
+  profile: ProfileInfo;
+  result: TAuthResult;
+};
+
+// @public
+export type SignInResolver<TAuthResult> = (
+  info: SignInInfo<TAuthResult>,
+  context: AuthResolverContext,
+) => Promise<BackstageSignInResult>;
+
+// @public (undocumented)
+export type StateEncoder = (req: OAuthStartRequest) => Promise<{
+  encodedState: string;
+}>;
+
+// @public
+export type TokenParams = {
+  claims: {
+    sub: string;
+    ent?: string[];
+  } & Record<string, JsonValue>;
 };
 ```
