+++ plugins/auth-node/src/alpha.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,20 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { createExtensionPoint } from '@backstage/backend-plugin-api';
-import { ProviderFactories } from '@backstage/plugin-auth-backend';
 
-/**
- * @alpha
- */
-export interface AuthProviderExtensionPoint {
-  addProviderFactories(factories: ProviderFactories): void;
-}
-
-/**
- * @alpha
- */
-export const authProviderExtensionPoint =
-  createExtensionPoint<AuthProviderExtensionPoint>({
-    id: 'auth.provider',
-  });
+export type { AuthProviderExtensionPoint } from './extensions';
+export { authProviderExtensionPoint } from './extensions';
